+++ src/macros/d6-roll.js
@@ -1,0 +1,1 @@
+new Roll('1d6').toMessage();