+++ src/macros/oracle-roll.js
@@ -1,0 +1,1 @@
+new Roll('1d100').toMessage();