+++ src/macros/starship-registry-number.js
@@ -1,0 +1,5 @@
+const letters = await fromUuid("Compendium.starsmith-expanded-oracles.starsmithexpandedoracles.RollTable.byBI6d46QtVH1goo");
+await letters.draw({ roll: true, displayChat: true });
+new Roll('1d10000-1').toMessage();
+const model = await fromUuid("Compendium.starsmith-expanded-oracles.starsmithexpandedoracles.RollTable.ZlVBOz7v2GzEuXtW");
+await model.draw({ roll: true, displayChat: true });